Move `RepositorySettings` out of Internal namespace
Make abstract
