+++ src/Tgstation.Server.Api/Models/ErrorCode.cs
@@ -185,7 +185,7 @@ namespace Tgstation.Server.Api.Models
 		SwarmIntegrityCheckFailed,
 
 		/// <summary>
-		/// One of <see cref="Internal.RepositorySettings.AccessUser"/> and <see cref="Internal.RepositorySettings.AccessToken"/> is set while the other isn't.
+		/// One of <see cref="RepositorySettings.AccessUser"/> and <see cref="RepositorySettings.AccessToken"/> is set while the other isn't.
 		/// </summary>
 		[Description("Either both accessUser and accessToken must be set or neither!")]
 		RepoMismatchUserAndAccessToken,
@@ -239,13 +239,13 @@ namespace Tgstation.Server.Api.Models
 		RepoDuplicateTestMerge,
 
 		/// <summary>
-		/// Attempted to set a whitespace <see cref="Internal.RepositorySettings.CommitterName"/>.
+		/// Attempted to set a whitespace <see cref="RepositorySettings.CommitterName"/>.
 		/// </summary>
 		[Description("committerName cannot be whitespace!")]
 		RepoWhitespaceCommitterName,
 
 		/// <summary>
-		/// Attempted to set a whitespace <see cref="Internal.RepositorySettings.CommitterEmail"/>.
+		/// Attempted to set a whitespace <see cref="RepositorySettings.CommitterEmail"/>.
 		/// </summary>
 		[Description("committerEmail cannot be whitespace!")]
 		RepoWhitespaceCommitterEmail,
@@ -377,19 +377,19 @@ namespace Tgstation.Server.Api.Models
 		InstanceMissingDreamMakerSettings,
 
 		/// <summary>
-		/// Missing <see cref="Internal.RepositorySettings"/> in database.
+		/// Missing <see cref="RepositorySettings"/> in database.
 		/// </summary>
 		[Description("Could not retrieve Repository settings from the database!")]
 		InstanceMissingRepositorySettings,
 
 		/// <summary>
-		/// Performing an automatic update with the <see cref="Internal.RepositorySettings.AutoUpdatesKeepTestMerges"/> flag resulted in merge conflicts.
+		/// Performing an automatic update with the <see cref="RepositorySettings.AutoUpdatesKeepTestMerges"/> flag resulted in merge conflicts.
 		/// </summary>
 		[Description("Performing this automatic update as a merge would result in conficts. Aborting!")]
 		InstanceUpdateTestMergeConflict,
 
 		/// <summary>
-		/// <see cref="Internal.RepositorySettings.AccessUser"/> and <see cref="Internal.RepositorySettings.AccessToken"/> are required for this operation.
+		/// <see cref="RepositorySettings.AccessUser"/> and <see cref="RepositorySettings.AccessToken"/> are required for this operation.
 		/// </summary>
 		[Description("Git credentials are required for this operation!")]
 		RepoCredentialsRequired,
