+++ src/Tgstation.Server.Host/Components/Repository/GitLabRemoteFeatures.cs
@@ -6,7 +6,6 @@ using GitLabApiClient;
 using Microsoft.Extensions.Logging;
 
 using Tgstation.Server.Api.Models;
-using Tgstation.Server.Api.Models.Internal;
 
 namespace Tgstation.Server.Host.Components.Repository
 {
