+++ src/Tgstation.Server.Host/Utils/GitHub/GitHubClientFactory.cs
@@ -13,7 +13,7 @@ using Microsoft.IdentityModel.Tokens;
 
 using Octokit;
 
-using Tgstation.Server.Api.Models.Internal;
+using Tgstation.Server.Api.Models;
 using Tgstation.Server.Host.Configuration;
 using Tgstation.Server.Host.System;
